Add Nebari class
