--- src/_nebari/__main__.py
@@ -1,10 +1,0 @@
-from _nebari.cli import create_cli
-
-
-def main():
-    cli = create_cli()
-    cli()
-
-
-if __name__ == "__main__":
-    main()
